Update ObSC.py
Sources are now dynamically listed with their address.  Instead of only being able to use /Scene #0-127, now /source/<sceneName> is accepted.  This is printed to the commandline.  getCurrentScenes will update the scenes from OBS and print them to the commandline.  Some work has been done to do a similar thing to sources, but not much has been done.  When parsing the Scene names from OBS, it will identify spaces in the Scene name and print out a warning, as spaces aren't valid OSC addresses.